backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
celery>=5.3
redis>=5.0
openpyxl>=3.1
geopy>=2.4
gunicorn>=21.2
python-dotenv>=1.0
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
